restructure of the game file
