+++ pokemaster/game/__init__.py
@@ -1,0 +1,2 @@
+from .game import Game
+from .type import TypeGame
